refactor(utils): co-locate listObjects module with consumers
Move listObjectsClient, filterObjects, and search (+tests) into a
single app/utils/listObjects/ directory. search.ts was a pure matcher
misfiled under components/GlobalSearch. Update the three connection-util
importers.
