+++ app/routes/objects/__tests__/objects.loader.test.ts
@@ -17,7 +17,7 @@ vi.mock("~/routes/connections/connections.server", () => ({
 }));
 
 const listObjectsClient = vi.fn();
-vi.mock("~/utils/listObjectsClient", () => ({
+vi.mock("~/utils/listObjects/listObjectsClient", () => ({
   listObjectsClient: (...args: unknown[]) => listObjectsClient(...args),
 }));
